backend/requirements.txt
flask
flask-cors
pyserial 